feat: refactor members content directory
empty files are handled gracefully as well, thank you bao for your help
