+++ src/content/members/2025.yaml
@@ -5,7 +5,7 @@
 - name: Alex Dediu
   short_bio: Hey! I am a first year computer science major.
   github: adediu25
-  website: alexandrudediu.weebly.com
+  website: https://alexandrudediu.weebly.com
 
 - name: Dylan Francis
   short_bio: Hi. I'm a first year computer science major.
