+++ src/content/members/2029.yaml
@@ -3,7 +3,7 @@
 
 - name: Daniel Budke
   github: dabudke
-  email: budkedp@mail.uc.edu
+  email: dbudke1010@outlook.com
   website: https://dabudke.github.io/
   linkedin: daniel-budke
   short_bio: >
