+++ src/content/members/alumni/2021.yaml
@@ -24,7 +24,7 @@
   short_bio: I'm a Computer Science student who enjoys python far too much for my own good.
   image_url: https://avatars1.githubusercontent.com/u/3240136
   github: nicholas-maltbie
-  website: nickmaltbie.com
+  website: https://nickmaltbie.com
 - name: Robert Bolgzds
   short_bio: Second year computer science major.
   image_url: https://avatars3.githubusercontent.com/u/31973394
